refactor(product_type): extract API layer into dojo/product_type/api/ [Phase 6,8,9]
Phase 6/8/9 of module reorg per AGENTS.md (Phase 7 N/A - no product_type API filter).
Move ProductTypeSerializer into api/serializer.py (re-exported from api_v2/serializers.py,
still used by ReportGenerateSerializer) and ProductTypeViewSet into api/views.py. Add
api/urls.py with add_product_type_urls() preserving route 'product_types' + basename
'product_type'; wire it into dojo/urls.py. Viewset re-export omitted (would cycle
api_v2.views<->product_type.api.views; only consumer was a test, now imports new path).
