+++ dojo/product_type/api/__init__.py
@@ -1,0 +1,1 @@
+path = "product_types"  # noqa: RUF067
